+++ Flow/src/utopia/flow/async/SynchronizedLoops.scala
@@ -3,7 +3,7 @@ package utopia.flow.async
 import utopia.flow.util.CollectionExtensions.RichTry
 import utopia.flow.time.WaitTarget
 import utopia.flow.time.WaitTarget.Until
-import utopia.flow.util.RichComparable._
+import utopia.flow.time.TimeExtensions._
 
 import scala.util.Try